+++ backend/src/Cowork.Infrastructure/Persistence/Configurations/ReservationConfiguration.cs
@@ -42,7 +42,7 @@ public sealed class ReservationConfiguration : IEntityTypeConfiguration<Reservat
 
         builder.Property(x => x.Status)
             .HasColumnName("status_id")
-            .HasConversion<int>()
+            .HasConversion<short>()
             .IsRequired();
 
         builder.Property(x => x.BaseAmount)
